requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
altair==5.1.2
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apache-superset==3.0.2
apispec==6.3.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.6.3
Bottleneck==1.3.7
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
constants==0.6.0
contourpy==1.1.1
convertdate==2.4.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==39.0.2
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
email-validator==1.3.1
executing==2.0.0
filelock==3.12.4
financedatabase==2.2.0
financetoolkit==1.5.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2023.9.2
func-timeout==4.3.5
gensim==4.3.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
h11==0.14.0
hashids==1.3.1
hijri-converter==2.3.1
holidays==0.23
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanize==4.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kombu==5.3.4
korean-lunar-calendar==0.3.1
limits==3.7.0
llvmlite==0.41.1
lxml==4.9.3
lyricsgenius==3.0.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
mutagen==1.47.0
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.14
numba==0.58.1
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
openai==0.28.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.1
pandas==1.5.3
paramiko==3.3.1
parsedatetime==2.6
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pgsanity==0.2.9
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.18.0
pocketsphinx==5.0.2
polyline==2.0.1
prison==0.2.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
PyMeeus==0.5.12
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-geohash==0.8.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.9.1
shortid==0.1.2
simplejson==3.19.2
six==1.16.0
sklearn==0.0.post10
slack-sdk==3.26.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
spotipy==2.23.0
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.4
srt==3.5.3
sshtunnel==0.4.0
stack-data==0.6.3
streamlit==1.27.2
sympy==1.12
tabulate==0.8.10
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.33.2
trio==0.23.1
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.5
validators==0.22.0
vercel==0.2.1
vine==5.1.0
vosk==0.3.44
wavio==0.0.7
wcwidth==0.2.8
webencodings==0.5.1
websockets==12.0
Werkzeug==2.3.8
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.1
WTForms-JSON==0.3.5
xarray==2023.10.1
XlsxWriter==3.0.9
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.31
yt-dlp==2023.12.30
zipp==3.17.0
